+++ src/sigobj.rs
@@ -9,6 +9,7 @@ use untrusted::Input;
 use super::x509::update_once;
 use super::cert::{Cert, ResourceCert};
 use super::x509::{Time, ValidationError};
+use signing::{DigestAlgorithm, SignatureAlgorithm};
 
 
 //------------ SignedObject --------------------------------------------------
@@ -219,56 +220,6 @@ impl SignedObject {
 }
 
 
-//------------ DigestAlgorithm -----------------------------------------------
-
-#[derive(Clone, Debug)]
-pub enum DigestAlgorithm {
-    Sha256,
-}
-
-
-impl DigestAlgorithm {
-    pub fn take_from<S: decode::Source>(
-        cons: &mut decode::Constructed<S>
-    ) -> Result<Self, S::Err> {
-        cons.take_sequence(Self::take_content_from)
-    }
-
-    pub fn take_opt_from<S: decode::Source>(
-        cons: &mut decode::Constructed<S>
-    ) -> Result<Option<Self>, S::Err> {
-        cons.take_opt_sequence(Self::take_content_from)
-    }
-
-    fn take_content_from<S: decode::Source>(
-        cons: &mut decode::Constructed<S>
-    ) -> Result<Self, S::Err> {
-        oid::SHA256.skip_if(cons)?;
-        cons.take_opt_null()?;
-        Ok(DigestAlgorithm::Sha256)
-    }
-
-
-    /// Parses a SET OF DigestAlgorithmIdentifiers.
-    ///
-    /// This is used in the digestAlgorithms field of the SignedData
-    /// container. It provides all the digest algorithms used later on, so
-    /// that the data can be read over. We don’t really need this, so this
-    /// function returns `()` on success.
-    ///
-    /// Section 2.1.2. of RFC 6488 requires there to be exactly one element
-    /// chosen from the allowed values.
-    pub fn skip_set<S: decode::Source>(
-        cons: &mut decode::Constructed<S>
-    ) -> Result<(), S::Err> {
-        cons.take_constructed_if(Tag::SET, |cons| {
-            while let Some(_) = Self::take_opt_from(cons)? { }
-            Ok(())
-        })
-    }
-}
-
-
 //------------ SignerInfo ----------------------------------------------------
 
 #[derive(Clone, Debug)]
@@ -441,35 +392,6 @@ impl SignedAttributes {
 }
 
 
-//------------ SignatureAlgorithm --------------------------------------------
-
-#[derive(Clone, Debug)]
-pub enum SignatureAlgorithm {
-    Sha256WithRsaEncryption
-}
-
-impl SignatureAlgorithm {
-    pub fn take_from<S: decode::Source>(
-        cons: &mut decode::Constructed<S>
-    ) -> Result<Self, S::Err> {
-        cons.take_sequence(Self::take_content_from)
-    }
-
-    pub fn take_content_from<S: decode::Source>(
-        cons: &mut decode::Constructed<S>
-    ) -> Result<Self, S::Err> {
-        let oid = Oid::take_from(cons)?;
-        if oid != oid::RSA_ENCRYPTION &&
-            oid != oid::SHA256_WITH_RSA_ENCRYPTION
-        {
-            return Err(decode::Malformed.into())
-        }
-        cons.take_opt_null()?;
-        Ok(SignatureAlgorithm::Sha256WithRsaEncryption)
-    }
-}
-
-
 //------------ OIDs ----------------------------------------------------------
 
 pub mod oid {
@@ -477,12 +399,6 @@ pub mod oid {
 
     pub const SIGNED_DATA: Oid<&[u8]>
         = Oid(&[42, 134, 72, 134, 247, 13, 1, 7, 2]);
-    pub const SHA256: Oid<&[u8]> = Oid(&[96, 134, 72, 1, 101, 3, 4, 2, 1]);
-    pub const RSA_ENCRYPTION: Oid<&[u8]>
-        = Oid(&[42, 134, 72, 134, 247, 13, 1, 1, 1]);
-    pub const SHA256_WITH_RSA_ENCRYPTION: Oid<&[u8]>
-        = Oid(&[42, 134, 72, 134, 247, 13, 1, 1, 11]);
-
     pub const CONTENT_TYPE: Oid<&[u8]>
         = Oid(&[42, 134, 72, 134, 247, 13, 1, 9, 3]);
     pub const MESSAGE_DIGEST: Oid<&[u8]>
